+++ instant/system/controllers/nordicblocks/backend/actions/catalog.php
@@ -1,0 +1,8 @@
+<?php
+
+class actionNordicblocksCatalog extends cmsAction {
+
+    public function run() {
+        return $this->redirect(href_to($this->controller->root_url, 'blocks'));
+    }
+}